requirements.txt
pyinstaller
pillow
tkinterdnd2
requests